+++ changes.md
@@ -5,6 +5,7 @@ DartMob Changes
 
 * Implemented Image. Added example.
 * Implemented FlowPanel. Added example.
+* Implemented HtmlPanel. Added example.
 
 **December 31, 2012